refactor: . Move createStatusRegistryReport() to new file StatusRegistryReport.ts
